pyproject.toml
[project]
name = "gymhelper"
version = "0.1.0"
description = "A FastAPI application for gym helper"
authors = [
    {name = "Kirill"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.11",
    "uvicorn>=0.30.1",
    "sqlalchemy>=2.0.31",
    "asyncpg>=0.29.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose>=3.3.0",
    "pydantic-settings>=2.5.2",
    "pyjwt>=2.9.0",
    "python-multipart>=0.0.12",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"